sync

+++ src/sccz80/plunge.c
@@ -323,16 +323,17 @@ void plnge2a(int (*heir)(LVALUE* lval), LVALUE* lval, LVALUE* lval2, void (*oper
             int doconstoper = 0;
             int32_t const_val;
 
+
             if ( lval2->is_const && (lval->val_type == KIND_INT || lval->val_type == KIND_CHAR || lval->val_type == KIND_LONG) ) {
                 doconstoper = 1;
-                const_val = lval2->const_val;
+                const_val = (uint32_t)lval2->const_val;
                 clearstage(before, 0);
                 force(rhs_val_type, lhs_val_type, lval->ltype->isunsigned, lval2->ltype->isunsigned, 0);
             }
             /* Handle the case that the constant was on the left */
             if ( lval1_wasconst && (lval2->val_type == KIND_INT || lval2->val_type == KIND_CHAR || lval2->val_type == KIND_LONG) ) {
                 doconstoper = 1;
-                const_val = lval->const_val;
+                const_val = (uint32_t)lval->const_val;
                 clearstage(before_constlval, 0);
                 force(lhs_val_type, rhs_val_type, lval2->ltype->isunsigned, lval2->ltype->isunsigned,0);
             }
